+++ packages/forma-36-react-components/src/components/DateTime/index.ts
@@ -1,0 +1,8 @@
+import { DateTime } from './DateTime/DateTime';
+import { RelativeDate } from './RelativeDate/RelativeDate';
+import * as datetime from './dateUtils';
+export {
+  datetime,
+  DateTime,
+  RelativeDate
+}