+++ packages/forma-36-react-components/src/components/Tabs/Tab.tsx
@@ -8,7 +8,7 @@ import classNames from 'classnames';
 
 import styles from './Tabs.css';
 
-export type TabProps = {
+export interface TabProps {
   id: string;
   onSelect?: (id: string, e: React.SyntheticEvent) => void;
   selected?: boolean;
@@ -20,9 +20,9 @@ export type TabProps = {
   className?: string;
   testId?: string;
   children: React.ReactNode;
-} & typeof defaultProps;
+}
 
-const defaultProps = {
+const defaultProps: Partial<TabProps> = {
   selected: false,
   disabled: false,
   testId: 'cf-ui-tab',
@@ -32,7 +32,7 @@ const defaultProps = {
 export class Tab extends Component<TabProps> {
   static defaultProps = defaultProps;
 
-  onClick: MouseEventHandler = e => {
+  onClick: MouseEventHandler = (e) => {
     if (this.props.onSelect && !this.props.disabled) {
       this.props.onSelect(this.props.id, e);
     }